requirements.txt
Flask==2.3.2
pymongo==4.4.1
python-dotenv==1.0.0
certifi==2023.7.22
PyJWT==2.8.0
bcrypt==4.0.1
requests==2.31.0
gunicorn==21.2.0